Initial Code commit
This is my initial commit with all of my code I have been working on. I
have an sql database containing two tables: Users which contains
information about all the users and Friends which contains pairs of
userIDs which denote one way friendship between two users. I also
included a encoder subsystem which exists in both the client and server
which is used to concatenate messages together for sending between the
client and server and between two clients. There is code for very basic
connectivity. The client is currently able to send login requests, and
logout requests. Any encryption has yet to be implemented at the moment.