Most functions work unsecured
Right now the server may run, and listen for connections, although it
works in a blocking format which is not ideal. It handles requests that
it receives by interacting with the database. Not all functionality has
been wired up yet, but logging in, logging out, requesting a friends
list (now included with logging in) and requesting contact information
all work. My next step is to implement this with a library to allow for
secure connectivity.